Refactor creation and freeze systems; update entity timing
Refactored OnCreation and Freeze systems for both server and shared engine, introducing OnCreationServerSystem and FreezeServerSystem for server-specific logic, and shared FreezeSystem and OnCreationSystem for general use. Changed create_time in EntityType and related classes from long to float for consistency. Updated InstructionManager and ComponentFactory to match new component signatures. Replaced shapeToCell/cellToShape with worldToCell/cellToWorld in Utility and RangedAttackSystem. Removed unused HealthSystem. Improved handling of entity creation, freeze, and position updates.

+++ android/src/main/java/io/github/android/gui/Card.java
@@ -16,7 +16,7 @@ public class Card {
     private final float passiveHeal;
     private final float baseSpeed;
     private final float freezeTime;
-    private final long createTime;
+    private final float createTime;
     private final HashMap<ResourcesType, Integer> cost;
 
     public Card(EntityType entityType, int imageResId) {
@@ -43,7 +43,7 @@ public class Card {
     public float getPassiveHeal() { return passiveHeal; }
     public float getBaseSpeed() { return baseSpeed; }
     public float getFreezeTime() { return freezeTime; }
-    public long getCreateTime() { return createTime; }
+    public float getCreateTime() { return createTime; }
     public HashMap<ResourcesType, Integer> getCost() { return cost; }
 
 }
